LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/Create-blocked-port.any-expected.txt

CONSOLE MESSAGE: WebSocket port 1 blocked
CONSOLE MESSAGE: WebSocket port 7 blocked
CONSOLE MESSAGE: WebSocket port 9 blocked
CONSOLE MESSAGE: WebSocket port 11 blocked
CONSOLE MESSAGE: WebSocket port 13 blocked
CONSOLE MESSAGE: WebSocket port 15 blocked
CONSOLE MESSAGE: WebSocket port 17 blocked
CONSOLE MESSAGE: WebSocket port 19 blocked
CONSOLE MESSAGE: WebSocket port 20 blocked
CONSOLE MESSAGE: WebSocket port 21 blocked
CONSOLE MESSAGE: WebSocket port 22 blocked
CONSOLE MESSAGE: WebSocket port 23 blocked
CONSOLE MESSAGE: WebSocket port 25 blocked
CONSOLE MESSAGE: WebSocket port 37 blocked
CONSOLE MESSAGE: WebSocket port 42 blocked
CONSOLE MESSAGE: WebSocket port 43 blocked
CONSOLE MESSAGE: WebSocket port 53 blocked
CONSOLE MESSAGE: WebSocket port 69 blocked
CONSOLE MESSAGE: WebSocket port 77 blocked
CONSOLE MESSAGE: WebSocket port 79 blocked
CONSOLE MESSAGE: WebSocket port 87 blocked
CONSOLE MESSAGE: WebSocket port 95 blocked
CONSOLE MESSAGE: WebSocket port 101 blocked
CONSOLE MESSAGE: WebSocket port 102 blocked
CONSOLE MESSAGE: WebSocket port 103 blocked
CONSOLE MESSAGE: WebSocket port 104 blocked
CONSOLE MESSAGE: WebSocket port 109 blocked
CONSOLE MESSAGE: WebSocket port 110 blocked
CONSOLE MESSAGE: WebSocket port 111 blocked
CONSOLE MESSAGE: WebSocket port 113 blocked
CONSOLE MESSAGE: WebSocket port 115 blocked
CONSOLE MESSAGE: WebSocket port 117 blocked
CONSOLE MESSAGE: WebSocket port 119 blocked
CONSOLE MESSAGE: WebSocket port 123 blocked
CONSOLE MESSAGE: WebSocket port 135 blocked
CONSOLE MESSAGE: WebSocket port 137 blocked
CONSOLE MESSAGE: WebSocket port 139 blocked
CONSOLE MESSAGE: WebSocket port 143 blocked
CONSOLE MESSAGE: WebSocket port 179 blocked
CONSOLE MESSAGE: WebSocket port 389 blocked
CONSOLE MESSAGE: WebSocket port 427 blocked
CONSOLE MESSAGE: WebSocket port 465 blocked
CONSOLE MESSAGE: WebSocket port 512 blocked
CONSOLE MESSAGE: WebSocket port 513 blocked
CONSOLE MESSAGE: WebSocket port 514 blocked
CONSOLE MESSAGE: WebSocket port 515 blocked
CONSOLE MESSAGE: WebSocket port 526 blocked
CONSOLE MESSAGE: WebSocket port 530 blocked
CONSOLE MESSAGE: WebSocket port 531 blocked
CONSOLE MESSAGE: WebSocket port 532 blocked
CONSOLE MESSAGE: WebSocket port 540 blocked
CONSOLE MESSAGE: WebSocket port 548 blocked
CONSOLE MESSAGE: WebSocket port 554 blocked
CONSOLE MESSAGE: WebSocket port 556 blocked
CONSOLE MESSAGE: WebSocket port 563 blocked
CONSOLE MESSAGE: WebSocket port 587 blocked
CONSOLE MESSAGE: WebSocket port 601 blocked
CONSOLE MESSAGE: WebSocket port 636 blocked
CONSOLE MESSAGE: WebSocket port 989 blocked
CONSOLE MESSAGE: WebSocket port 990 blocked
CONSOLE MESSAGE: WebSocket port 993 blocked
CONSOLE MESSAGE: WebSocket port 995 blocked
CONSOLE MESSAGE: WebSocket port 1719 blocked
CONSOLE MESSAGE: WebSocket port 1720 blocked
CONSOLE MESSAGE: WebSocket port 1723 blocked
CONSOLE MESSAGE: WebSocket port 2049 blocked
CONSOLE MESSAGE: WebSocket port 3659 blocked
CONSOLE MESSAGE: WebSocket port 4045 blocked
CONSOLE MESSAGE: WebSocket port 6000 blocked
CONSOLE MESSAGE: WebSocket port 6566 blocked
CONSOLE MESSAGE: WebSocket port 6665 blocked
CONSOLE MESSAGE: WebSocket port 6666 blocked
CONSOLE MESSAGE: WebSocket port 6667 blocked
CONSOLE MESSAGE: WebSocket port 6668 blocked
CONSOLE MESSAGE: WebSocket port 6669 blocked
CONSOLE MESSAGE: WebSocket port 6697 blocked
CONSOLE MESSAGE: WebSocket port 10080 blocked

PASS Basic check
PASS WebSocket blocked port test 1
PASS WebSocket blocked port test 7
PASS WebSocket blocked port test 9
PASS WebSocket blocked port test 11
PASS WebSocket blocked port test 13
PASS WebSocket blocked port test 15
PASS WebSocket blocked port test 17
PASS WebSocket blocked port test 19
PASS WebSocket blocked port test 20
PASS WebSocket blocked port test 21
PASS WebSocket blocked port test 22
PASS WebSocket blocked port test 23
PASS WebSocket blocked port test 25
PASS WebSocket blocked port test 37
PASS WebSocket blocked port test 42
PASS WebSocket blocked port test 43
PASS WebSocket blocked port test 53
PASS WebSocket blocked port test 69
PASS WebSocket blocked port test 77
PASS WebSocket blocked port test 79
PASS WebSocket blocked port test 87
PASS WebSocket blocked port test 95
PASS WebSocket blocked port test 101
PASS WebSocket blocked port test 102
PASS WebSocket blocked port test 103
PASS WebSocket blocked port test 104
PASS WebSocket blocked port test 109
PASS WebSocket blocked port test 110
PASS WebSocket blocked port test 111
PASS WebSocket blocked port test 113
PASS WebSocket blocked port test 115
PASS WebSocket blocked port test 117
PASS WebSocket blocked port test 119
PASS WebSocket blocked port test 123
PASS WebSocket blocked port test 135
PASS WebSocket blocked port test 137
PASS WebSocket blocked port test 139
PASS WebSocket blocked port test 143
PASS WebSocket blocked port test 179
PASS WebSocket blocked port test 389
PASS WebSocket blocked port test 427
PASS WebSocket blocked port test 465
PASS WebSocket blocked port test 512
PASS WebSocket blocked port test 513
PASS WebSocket blocked port test 514
PASS WebSocket blocked port test 515
PASS WebSocket blocked port test 526
PASS WebSocket blocked port test 530
PASS WebSocket blocked port test 531
PASS WebSocket blocked port test 532
PASS WebSocket blocked port test 540
PASS WebSocket blocked port test 548
PASS WebSocket blocked port test 554
PASS WebSocket blocked port test 556
PASS WebSocket blocked port test 563
PASS WebSocket blocked port test 587
PASS WebSocket blocked port test 601
PASS WebSocket blocked port test 636
PASS WebSocket blocked port test 989
PASS WebSocket blocked port test 990
PASS WebSocket blocked port test 993
PASS WebSocket blocked port test 995
PASS WebSocket blocked port test 1719
PASS WebSocket blocked port test 1720
PASS WebSocket blocked port test 1723
PASS WebSocket blocked port test 2049
PASS WebSocket blocked port test 3659
PASS WebSocket blocked port test 4045
PASS WebSocket blocked port test 6000
PASS WebSocket blocked port test 6566
PASS WebSocket blocked port test 6665
PASS WebSocket blocked port test 6666
PASS WebSocket blocked port test 6667
PASS WebSocket blocked port test 6668
PASS WebSocket blocked port test 6669
PASS WebSocket blocked port test 6697
PASS WebSocket blocked port test 10080
